Updated ESMF dockerfile
This patch updates the base OS for ESMF (and in turn all the other
images that depend on it) from Debian stretch to Debian buster. This
also requires a more recent version of ESMF than the current release
(7.1.0) to fix compilation errors with gcc-8. The commit hash for ESMF
has been hard-coded into the dockerfile for reproducibility.